Move isVSCodeExtension func to avoid circular dependency
