+++ src/emulator/hub.ts
@@ -11,7 +11,7 @@ import { EmulatorRegistry } from "./registry";
 import { FunctionsEmulator } from "./functionsEmulator";
 import { ExpressBasedEmulator } from "./ExpressBasedEmulator";
 import { PortName } from "./portUtils";
-import { isVSCodeExtension } from "../utils";
+import { isVSCodeExtension } from "../vsCodeUtils";
 
 // We use the CLI version from package.json
 const pkg = require("../../package.json");
